refactor: match components and function import
