+++ app/layout.tsx
@@ -1,7 +1,7 @@
+import "./globals.css";
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
-import "./globals.css";
-import Navbar from "./components/Navbar";
+import { Navbar } from "../components/common/Navbar";
 import { SpeedInsights } from "@vercel/speed-insights/next"
 
 const geistSans = Geist({
